Move stack with limit utils to runner.rs
